Fix 'Forbidden' error for users on first login
